//Почти сделал сепокку
Поменял области видимости и изменил BaseHttpHandler и его наследников согласно рекомендациям
