+++ src/manager/http/HttpTaskServer.java
@@ -17,7 +17,7 @@ public class HttpTaskServer {
 
     protected Gson gson;
 
-    protected TaskManager taskManager;
+    private final TaskManager taskManager;
     protected HttpServer server;
     private static final int PORT = 8080;
