+++ src/manager/http/handlers/SubTaskHandlers.java
@@ -14,11 +14,9 @@ public class SubTaskHandlers extends BaseHttpHandler {
     }
 
     @Override
-    public void handle(HttpExchange exchange) throws IOException {
+    public void safeHandle(HttpExchange exchange) throws IOException {
         Endpoint endpoint = getEndpoint(exchange);
         String[] split = exchange.getRequestURI().getPath().split("/");
-
-        try {
             switch (endpoint) {
                 case GET ->
                     sendText(exchange, gson.toJson(taskManager.showAllSubTasks()));
@@ -42,12 +40,5 @@ public class SubTaskHandlers extends BaseHttpHandler {
                 case UNKNOWN ->
                     sendNotFound(exchange);
             }
-        } catch (NullPointerException e) {
-            sendNotFound(exchange);
-        } catch (ManagerSaveException e) {
-            sendHasInteractions(exchange);
-        } catch (Exception e) {
-            writeResponse(exchange, 500, "");
         }
     }
-}
